backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pandas==2.2.3
openpyxl==3.1.5
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.1
psycopg2-binary==2.9.9
email-validator==2.2.0
requests